Added color args
Added color Args and changed the printing of the background colors
to only happen when given other wise the default is users background.

Added run time controls.
Up and Down Moves threw list of figlet fonts.
Right and Left change the size of the output width of figlet.

Code clean up:
    All function now return a `Box<dyn std::error::Error>>` over
    what we was returning the 'crossterm::Result' type.
    This always us to have more error type to use with the `?`.

Added a link to more figlet fonts in the Readme.